refactor: separate connection & file transfer functionality into custom hooks
